+++ packages/server/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@acs/shared": "1.0.0",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-validator": "^6.6.1",
@@ -20,6 +21,7 @@
     "node-cron": "^2.0.3"
   },
   "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
     "@types/cors": "^2.8.7",
     "@types/express": "^4.17.8",
     "@types/express-validator": "^3.0.0",
