+++ packages/server/src/models/types.ts
@@ -1,0 +1,5 @@
+export enum ModelRefs {
+  USER = 'User',
+  POSITION = 'Position',
+  NOMINATION = 'Nomination',
+}
