+++ packages/shared/src/models/nomination.ts
@@ -1,10 +1,9 @@
-import type { UserDoc } from './user';
-import type { PositionDoc } from './position';
+import type UserDoc from './user';
+import type PositionDoc from './position';
 
-import { Types, Document, Schema, model } from 'mongoose';
-import { ModelRefs } from './types';
+import { Types, Document } from 'mongoose';
 
-export interface NominationDoc extends Document {
+interface NominationDoc extends Document {
   position: Types.ObjectId | PositionDoc;
   candidate: Types.ObjectId | UserDoc;
   seconds: Types.Array<Types.ObjectId>;
@@ -12,36 +11,4 @@ export interface NominationDoc extends Document {
   writeUp?: string;
 }
 
-const NominationSchema = new Schema({
-  position: {
-    type: Types.ObjectId,
-    required: true,
-    ref: ModelRefs.POSITION,
-  },
-  candidate: {
-    type: Types.ObjectId,
-    required: true,
-    ref: ModelRefs.USER,
-  },
-  seconds: [
-    {
-      type: Types.ObjectId,
-      required: true,
-    },
-  ],
-  video: {
-    type: String,
-    trim: true,
-  },
-  writeUp: {
-    type: String,
-    trim: true,
-  },
-});
-
-const NominationModel = model<NominationDoc>(
-  ModelRefs.NOMINATION,
-  NominationSchema,
-);
-
-export default NominationModel;
+export default NominationDoc;
