+++ packages/shared/src/models/position.ts
@@ -1,38 +1,12 @@
-import type { UserDoc } from './user';
+import type UserDoc from './user';
 
-import { Types, Document, Schema, model } from 'mongoose';
-import { ModelRefs } from './types';
+import { Types, Document } from 'mongoose';
 
-export interface PositionDoc extends Document {
+interface PositionDoc extends Document {
   title: string;
   description: string;
   isOpen: boolean;
   occupant: Types.ObjectId | UserDoc;
 }
 
-const PositionSchema = new Schema({
-  title: {
-    type: String,
-    required: true,
-    trim: true,
-  },
-  description: {
-    type: String,
-    required: true,
-    trim: true,
-  },
-  isOpen: {
-    type: Boolean,
-    required: true,
-    default: false,
-  },
-  occupant: {
-    type: Types.ObjectId,
-    required: true,
-    ref: ModelRefs.USER,
-  },
-});
-
-const PositionModel = model<PositionDoc>(ModelRefs.POSITION, PositionSchema);
-
-export default PositionModel;
+export default PositionDoc;
